Revising kwargs and style settings of plot_data, fixes
Completes #1.
Making legpos and legcol a String/Int64 only not allowing a distinction
between different axes with adjustments to function setup_axes.
Ensuring one legend is printed in function plot_data, when a second axis
is defined.
Defining a new index for the colour scheme, which is continuously used
to define colours and line styles in case of 2 pairs of y axes, if only
one colour scheme is used and is reset to 1 for an alternative second
colour scheme.
Introducing aliases for kwargs ti, cs, lc, lt, pt, and legpos.
New functions checkaliases and checkalias to check correct assignment of
each kwargs without unintentional double assignments.
Fix: For type of field dashes in PlotData, use Union of
     Tuple{Float64,Float64} and Tuple{Int64,Int64} rather than
     Tuple{Number,Number}. Adjusted types for kwarg `lt` in function
     load_PlotData.
Fix: In function setup_axes, only transform cs to [cs, cs], when a
     second axis is defined and cs is not a vector.