Set colour schemes/format stack in plot_stack
At the beginning of plot_stack, make a deepcopy of the plot_list to not
overwrite values in each plot_data and have original PlotData of the use
of plot_stack.
Call of function set_style from plot_stack to format stack and/or set a
colour scheme.
Rename kwarg borders in plot_stack to border.