Fixes and Updates
In plot_stack and plot_data in public.jl, change default value of alpha
to 1 and adjust conditions where alpha must be >= 0 in plot.jl.
Remove function call of set_style in plot_stack and correct argument in
print_stack from lt to linetype. Reverse pltdata to [pltdata] in
argument list of set_axes and format_axes_and_annotations.
xlims can now only be a tuple or nothing and is set/checked in
setup_plot, otherwise an ArgumentError is thrown.
In print_stack, use lt[i], not lt, for dashes of borders to allow
different border styles.
In format_timeseries, hand over xlims, not xlims[1], as xlims is not an
array for each axis and otherwise only the lower limit gets handed over.
In datamanipulation.jl in index2range, set firstextra and lastextra to
empty ranges, when index is empty.